Add release workflow to build releases automatically
This adds a continuous delivery pipeline which will run on release tags.
It currently will build a release binary for Linux x86_64, Linux ARM,
and MacOS x86_64 (which can be run on M* series macs).

It has code in it already to support cross compiling using
[cross](https://github.com/cross-rs/cross) which can be setup later.
I've yet to get it working flawlessly on my own projects so I have left
the setup code in but not utilised it.

This fixes #50