refactor: better diffing
'Hunk' now stores formatted 'Text'.

There was an issue where what was shown and translated
into lines on the screen was the result of a
second diff between Hunks done by the 'similar' lib.

Now instead, 'similar' diffs entire files,
which makes it easier to correlate about what is
shown on screen, and what is being included in
a patch.

Moreover, this will make it easier to
apply code syntax-highlighting within diffs.