feat: FileWatcher now ignores changes from patterns in .gitignore
`FileWatcher` will build a gitignore ruleset from all `.gitignore` files
in the repo and any global `.gitignore`, and use this to ignore file
change events. This prevents a lot of `FileUpdate` events being
generated when ignored items like build folders change, which may have
been contributing to freezes.

If a change is detected involving a `.gitignore` file, the ruleset is rebuilt.

Changes are also now no longer watched from a thread, which caused problems with some watcher implementations such as `Fsevents` on macOS.