fix: report key bindings without "<>"
Fixes #256

This commit does a few things. Primarily, it validates that there are no key
bindings like "ctrl+g" (these should be "<ctrl+g>", i.e. surrounded by angle
brackets).

To ensure that the bindings are correct for the rest of the code, the `Config`
struct has been split into something that `figment` recognises and another
that's actually used.

Finally, if the whole program has an error, the `Display` representation is
printed, not the `Debug` one; this makes the error messages easier for users to
read.

It would be better to detect errors in the `nom` code, but I lack experience
with this and get a headache every single time :)