refactor: Split Ui/State under a new App struct
This makes it possible to render to screen with access to the
state.

before:
```
// Won't work, since state is borrowed mutably twice
state.ui.render_to_screen(&mut surface, &mut state);
```

after:
```
app.ui.render_to_screen(&mut surface, &mut app.state);
```

Previously the Ui was always created for every single frame with
Ratatui, so this split was not a requirement.
Updates of the App may still render to screen so they need access to Ui.