feat: optimize & defer rendering of items in editor, esp. diff hunk highlights (#392)
* draft(items): move away from `Line` and `Span`

* feat: JIT rendering

NOTE: we still need to implement the rendering of commits

* feat: implement commit rendering

* fix: typo in branch status text

* fix: restore replace of TAB with 4 spaces

* perf: avoid some allocations

* ref: `TargetData::Commit` now has named members

this makes more sense since we store a lot of different data

* ref: `TargetData` is now `ItemData`

* fix: construct lines again

* fix: adding some missing section headers

* fix: make `Item.data` is no longer optional

Something is very broken with diffing in this commit lol.

On the bright side; We are rendering the status items and commits again!

* chore: remove old comment

* fix: removing very hacky logging that made it thru

* fix: correctly highlight hunks again

* feat: `ItemData::to_line` -> `ItemData::to_lines`

this method now returns a `Vec<Line>` instead of a single `Line`

NOTE: hunk highlighting is crashing in this commit for some reason :(

* fix: (mostly) restore hunk line rendering

* chore: remove old comment

* fix: `ItemData::to_line` now produces a single line

* fix: `ItemData::Raw` holds an owned String

* fix: restore commit view header

* fix: restore status section headers

* fix: typo in stash string

* feat: cache syntax highlighting for hunks

* fix: unify hunk line range scanning into one fn

Every part of the hunk processing that requires
scanning hunk content to find line ranges now all
use `gitu_diff::line_range_iterator` instead of using
in-place solutions. This is to make line scanning
more consistent and more easily maintainable.

* chore: remove unused imports

* fix: store visual reference prefix to apply later

this should fix any issues introduced by prefixing
the reference shorthand when constructing the `ItemData`.

* chore: remove old comment

* fix: restore creation of `RefKind::Remote` items

* ref: skip replacing '\t' in hunks for now

* chore: update tests related to branch divergence

* fix: `Diff::hunk_content` to unify getting hunk content

* fix: restore stash item styling

* fix: generate more collision resistant cache-key

This seems to satisfy tests involving diffs for now.
Not sure if this is a preferrable solution but it should
suffice.

* chore: update branch divergence tests

* fix: replace tab-characters with 4 spaces again

* fix: restore ref styling

* fix: return commits from `selected_rev` again

this takes care of the last of the tests that were failing

* ref: cleanup create reference prefix function

* chore: `line_range_iterator` doc-comment update

* ref: resolve clippy warnings

* fix: use already computed hunk hash for caching

* ref: move `line_range_iterator` to src/highlight.rs

* chore: remove doc-comment

this fails compilation even though it shouldn't
so I just removed it lol

* ref: clean up parts of hunk line highlighting

* ref: impl `to_line` on `Item` directly instead

* fix: don't format!() hunk highlight cache-keys

* ref: move `Item::to_line` to `src/items.rs`

* fix: remotes and tags can't be deleted or discarded

* ref: abstract away section ids for maintainability