WIP feat: Report bad config key bindings to the user
WIP: tests fail and I'm not sure how this is supposed to work

This commit changes the `nom` key parsing code such that key bindings in a
user's config do not surround special keys (e.g. "ctrl") in angle brackets, and
any errors encountered when parsing these are immediately reported to the user.